#80 Refactored employee service exception handling
